further rationalize downloader
